edgedetection/requirements.txt
numpy
matplotlib
opencv-python
scipy
scikit-image
tqdm
Pillow
torch
torchvision
